Add PV/PVC cleanup verification to existing modelcache E2E tests
Move cleanup checks from a standalone test into the existing modelcache
E2E tests per review feedback. Each test now verifies that serving
PV/PVC resources are cleaned up after LocalModelCache deletion.

Shared helpers (assert_pv_deleted, assert_pvc_deleted) are added to
test/e2e/modelcache/__init__.py to avoid duplication.
