+++ test/e2e/modelcache/test_localmodelnamespacecache.py
@@ -44,6 +44,7 @@ from kserve.models.v1beta1_predictor_spec import V1beta1PredictorSpec
 from kserve.models.v1beta1_model_spec import V1beta1ModelSpec
 from kserve.models.v1beta1_model_format import V1beta1ModelFormat
 from ..common.utils import KSERVE_TEST_NAMESPACE, predict_isvc
+from . import assert_pv_deleted, assert_pvc_deleted
 
 
 @pytest.mark.modelcache
@@ -187,4 +188,16 @@ async def test_sklearn_modelnamespacecache(rest_v1_client, network_layer):
     kserve_client.delete_local_model_namespace_cache(
         model_cache.metadata.name, namespace=KSERVE_TEST_NAMESPACE
     )
+
+    # Verify PV/PVC are cleaned up after LocalModelNamespaceCache deletion.
+    # Namespace-scoped caches require explicit delete RBAC since PVs cannot have
+    # owner references to namespace-scoped resources.
+    core_api = client.CoreV1Api()
+    serving_pv = f"{model_cache.metadata.name}-{node_group.metadata.name}-{KSERVE_TEST_NAMESPACE}"
+    serving_pvc = f"{model_cache.metadata.name}-{node_group.metadata.name}"
+    download_pv = f"{model_cache.metadata.name}-{node_group.metadata.name}-{KSERVE_TEST_NAMESPACE}-download"
+    await assert_pv_deleted(core_api, serving_pv)
+    await assert_pvc_deleted(core_api, serving_pvc, KSERVE_TEST_NAMESPACE)
+    await assert_pv_deleted(core_api, download_pv)
+
     kserve_client.delete_local_model_node_group(node_group.metadata.name)
